python/requirements.txt
tabulate
tqdm
pyclair
kubernetes
PyYAML